restore nullish
