+++ crates/swc_ecma_compiler/src/features.rs
@@ -7,8 +7,7 @@ bitflags! {
         const OPTIONAL_CHAINING = 1 << 1;
         const PRIVATE_IN_OBJECT = 1 << 2;
         const LOGICAL_ASSIGNMENTS = 1 << 3;
-        const NULLISH_COALESCING = 1 << 4;
-        const EXPORT_NAMESPACE_FROM = 1 << 5;
+        const EXPORT_NAMESPACE_FROM = 1 << 4;
     }
 }
